Expedition Action Reducer

+++ client/src/expedTable/expedTableRow.js
@@ -1,18 +1,18 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { connect } from "react-redux";
-import { deleteExpedition} from "../redux/actions/allExped.actions";
+import { withRouter } from "react-router-dom";
+import { deleteExpedition } from "../redux/actions/allExped.actions";
 import { Button } from "reactstrap";
 
 const ExpedTableRow = props => {
   const { _id, numberExpedition, dateExpedition, caveName } = props.item;
 
-  const deleteItem = async _id => {
-    try {
-      return await props.delete(_id);
-    } catch (err) {
-      return console.log(err);
-    }
+  const deleteItem = _id => {
+    return props.delete(_id);
+  };
+  const updateItem = _id => {
+    return props.history.push(`/admin/update-expedition/${_id}`);
   };
   return (
     <tr>
@@ -24,15 +24,21 @@ const ExpedTableRow = props => {
       <td>{dateExpedition}</td>
       <td>{caveName}</td>
       <td>
-        <Button outline size="sm" color="primary">
+        <Button
+          onClick={() => {
+            updateItem(_id);
+          }}
+          outline
+          size="sm"
+          color="primary"
+        >
           Update
         </Button>{" "}
       </td>
       <td>
         <Button
           onClick={() => {
-            if (window.confirm("Are you sure you wish to delete this item?"))
-              deleteItem(_id);
+            if (window.confirm("Are you sure ?")) deleteItem(_id);
           }}
           outline
           size="sm"
@@ -49,7 +55,9 @@ const mapDispatchToProps = dispatch => {
     delete: id => dispatch(deleteExpedition(id))
   };
 };
-export default connect(
-  null,
-  mapDispatchToProps
-)(ExpedTableRow);
+export default withRouter(
+  connect(
+    null,
+    mapDispatchToProps
+  )(ExpedTableRow)
+);
